AWSAirflowDemo/requirements.txt
apache-airflow-backport-providers-amazon
apache-airflow-backport-providers-postgres
apache-airflow[jdbc]==1.10.12